pyproject.toml
[project]
name = "hafs"
version = "0.1.0"
description = "Halext Agentic File System - AFS context manager with TUI"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "scawful"}
]
keywords = ["tui", "afs", "context", "ai", "gemini", "claude", "halext"]

dependencies = [
    "textual>=0.95.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "ptyprocess>=0.7.0",
    "rapidfuzz>=3.0.0",
    "tomli-w>=1.0.0",
    "textual-terminal>=0.3.0",
    "pyte>=0.8.0",
    "python-ulid>=3.0.0",
    "aiohttp>=3.9.0",
    "numpy>=1.24.0",
    "plotext>=5.2.8",
    "psutil>=5.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
backends = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]
training-node = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "textual-dev>=1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
hafs = "hafs.cli:main"

[build-system]
requires = ["scikit-build-core>=0.5", "pybind11>=2.11"]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
cmake.minimum-version = "3.15"
wheel.packages = [
    "src/hafs",
    "src/agents",
    "src/backends",
    "src/services",
    "src/core",
    "src/config",
    "src/models",
    "src/synergy",
    "src/context",
    "src/tui",
    "src/cli",
    "src/llm",
    "src/adapters",
    "src/plugins",
    "src/integrations",
    "src/editors"
]
cmake.source-dir = "src/cc"
build-dir = "build/{wheel_tag}"

[tool.scikit-build.cmake.define]
CMAKE_OSX_ARCHITECTURES = "arm64"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.mypy-pytest]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
